refactor: use numeric types from stdint.h library

+++ atoms.h
@@ -1,7 +1,7 @@
 #ifndef ATOM
 #define ATOM
 
-#include <stddef.h>
+#include <stdint.h>
 
 typedef struct atomdesc {
   double x_pos;
@@ -13,10 +13,10 @@ typedef struct atom_list {
   double *x_pos;
   double *y_pos;
   double *z_pos;
-  size_t len;
+  uint64_t len;
 } atoms_data;
 
-void atoms_data_init(atoms_data *atom, unsigned int box_size);
+void atoms_data_init(atoms_data *atom, uint64_t box_size);
 void atoms_data_cleanup(atoms_data *atoms);
 
 #endif // !ATOM
